Handle uploads into a subfolder

+++ system/ee/ExpressionEngine/Model/File/UploadDestination.php
@@ -537,13 +537,11 @@ class UploadDestination extends StructureModel
         $filesystem = $this->getFilesystem();
 
         // Remove the file
-        if ($this->exists($path)) {
-            if ($this->isDirectory($path)) {
-                $filesystem->deleteDir($path);
-            } else {
-                $filesystem->delete($path);
-            }
+        if ($filesystem->exists($path)) {
+            return $filesystem->delete($path);
         }
+
+        return false;
     }
 
     public function deleteGeneratedFiles($path)
@@ -573,12 +571,14 @@ class UploadDestination extends StructureModel
         $basename = ($renamer === false) ? $basename : substr($basename, 0, -strlen($renamer));
 
         foreach ($manipulations as $manipulation) {
-            $files = $filesystem->getDirectoryContents("{$dirname}/_{$manipulation}/");
-            $files = array_filter($files, function ($file) use ($basename) {
-                return (strpos($file, "{$basename}_") === 0);
-            });
-            foreach ($files as $file) {
-                $filesystem->delete($file);
+            if($filesystem->exists("{$dirname}/_{$manipulation}/")) {
+                $files = $filesystem->getDirectoryContents("{$dirname}/_{$manipulation}/");
+                $files = array_filter($files, function ($file) use ($basename) {
+                    return (strpos($file, "{$basename}_") === 0);
+                });
+                foreach ($files as $file) {
+                    $filesystem->delete($file);
+                }
             }
         }
     }
